testcore: fix all remaining failing core tests for 5.14.4-nt
This testsuite was written by our cPanel code_monkeys team:
  Todd Rinaldo, Dan Muey, Nicolas Rochelemagne, Cleland Early

fix taint handling in subtests
fix qr/^\Q quoting and wrong B:: package in op/method.t
TODO t/CORE/op/write.t and skip op/method.t UNIVERSAL::AUTOLOAD
keep temp files with BC_DEVELOPING=1
TODO io/utf8.t
fix op/gv.t
untodo comp/proto.t, op/method.t, op/reset.t
add make testcore to master Makefile

5.18.2-nt has several errors and TODOs:

t/C-COMPILED/CORE--comp--bproto.t                (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  7-8, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--comp--retainedlines.t         (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--comp--parser.t                (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--comp--require.t               (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  7-8, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--io--open.t                    (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--array_base.t              (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--io--utf8==BADPLAN-255.t       (Wstat: 0 Tests: 13 Failed: 0)
  TODO passed:   8
t/C-COMPILED/CORE--op--assignwarn.t              (Wstat: 256 Tests: 13 Failed: 1)
  Failed test:  13
  Non-zero exit status: 1
t/C-COMPILED/CORE--op--bless.t                   (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--bop.t                     (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--closure==BADTEST-254.t    (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  7-8
  TODO passed:   10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--cproto.t                  (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--dbm==BADTEST-208-226.t    (Wstat: 0 Tests: 13 Failed: 0)
  TODO passed:   10-11
t/C-COMPILED/CORE--op--exists_sub==BADTEST-251.t (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  7-8
  TODO passed:   10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--die_keeperr.t             (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--filetest.t                (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--filetest_stack_ok.t       (Wstat: 256 Tests: 13 Failed: 1)
  Failed test:  13
  Non-zero exit status: 1
t/C-COMPILED/CORE--op--eval.t                    (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--gv.t                      (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  7-8, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--op--hash.t                    (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  8-9, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--op--hashassign.t              (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--leaky-magic.t             (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  8-9, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--op--lex_assign.t              (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--lc_user.t                 (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--lc.t                      (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  7-8, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--op--length.t                  (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--local.t                   (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--mydef.t                   (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  8-9, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--op--magic==BADTEST-281.t      (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  7-8
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--packagev==BADTEST-249.t   (Wstat: 0 Tests: 13 Failed: 0)
  TODO passed:   10-11
t/C-COMPILED/CORE--op--quotemeta.t               (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--range.t                   (Wstat: 256 Tests: 13 Failed: 1)
  Failed test:  13
  Non-zero exit status: 1
t/C-COMPILED/CORE--op--readdir.t                 (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--pack.t                    (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  8-9, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--op--smartkve.t                (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--sigdispatch.t             (Wstat: 256 Tests: 13 Failed: 1)
  Failed test:  3
  Non-zero exit status: 1
t/C-COMPILED/CORE--op--sort.t                    (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--split_unicode.t           (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  7-8, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--op--sub_lval.t                (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  10-11, 13
  Non-zero exit status: 3
t/C-COMPILED/CORE--op--tie.t                     (Wstat: 1024 Tests: 13 Failed: 4)
  Failed tests:  8, 10-12
  Non-zero exit status: 4
t/C-COMPILED/CORE--op--switch==BADTEST-180.t     (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  8-9
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--taint==BC-288.t           (Wstat: 0 Tests: 13 Failed: 0)
  TODO passed:   4-7, 13
t/C-COMPILED/CORE--op--undef.t                   (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  7-8, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--op--turkish.t                 (Wstat: 512 Tests: 13 Failed: 2)
  Failed tests:  10-11
  Non-zero exit status: 2
t/C-COMPILED/CORE--op--utf8decode.t              (Wstat: 256 Tests: 13 Failed: 1)
  Failed test:  13
  Non-zero exit status: 1
t/C-COMPILED/CORE--op--utftaint.t                (Wstat: 768 Tests: 13 Failed: 3)
  Failed tests:  8-9, 12
  Non-zero exit status: 3
t/C-COMPILED/CORE--op--write===BADTEST-238.t     (Wstat: 256 Tests: 13 Failed: 1)
  Failed test:  8
  Non-zero exit status: 1
Files=237, Tests=3081, 211 wallclock secs ( 3.63 usr  0.28 sys + 706.32 cusr 26.82 csys = 737.05 CPU)
Result: FAIL

+++ t/CORE/test.pl
@@ -702,7 +702,7 @@ sub unlink_all {
 }
 
 my %tmpfiles;
-END { unlink_all keys %tmpfiles }
+END { unlink_all keys %tmpfiles unless $ENV{BC_DEVELOPING}}
 
 # A regexp that matches the tempfile names
 $::tempfile_regexp = 'tmp\d+[A-Z][A-Z]?';
@@ -748,7 +748,7 @@ sub _fresh_perl {
         ($tmpfile) = $tmpfile =~ m/(.*)/;
         $tmpfile =~ s/\.bin$/.subtest.$test.t/;
         $tmpfiles{$tmpfile} = 1;
-        unlink $tmpfile if -e $tmpfile;
+        unlink $tmpfile if -e $tmpfile and !$ENV{BC_DEVELOPING};
     }
 
     # Given the choice of the mis-parsable {}
@@ -838,14 +838,18 @@ sub runperl_binary {
     my ( $test, $opts ) = @_;
 
     $opts ||= {};
+    #print STDERR @{$opts->{'switches'}},"\n";
     my $error = $opts->{'stderr'} ? '2>&1' : '';
+    my $taint = $opts->{'switches'} ? join(' ',grep /-[tT]/, @{$opts->{'switches'}}) : '';
     my $bin = $test;
     $bin =~ s/\.t$/\.bin/;
     unlink $bin if -e $bin;
-    print STDERR "# running: make $bin ===\n";
-
     ( $ENV{PATH} ) = $ENV{PATH} =~ m/(.*)/;
-    my $make = `blib/script/perlcc -O3 -o $bin $test $error`;
+    my $cmd = "$^X -Iblib/arch -Iblib/lib script/perlcc $taint -O3 -o $bin $test $error";
+    ( $cmd ) = $cmd =~ m/(.*)/;
+    print STDERR "# running: make $bin\n";
+    print STDERR "# $cmd\n" if $ENV{TEST_VERBOSE};
+    my $make = `$cmd`;
     map { print STDERR "# $_\n" } split /\n/, $make;
     return $make if $? || $opts->{perlcc_only};
 
@@ -859,7 +863,7 @@ sub runperl_binary {
     else {
         $output = `./$bin $error`;
     }
-    unlink $bin;
+    unlink $bin unless $ENV{BC_DEVELOPING};
     return $output;
 }
 
@@ -1057,11 +1061,13 @@ sub run_multiple_progs {
 
         ok($ok);
 
-        foreach (@temps) {
+        if (!$ENV{BC_DEVELOPING}) {
+          foreach (@temps) {
             unlink $_ if $_;
-        }
-        foreach (@temp_path) {
+          }
+          foreach (@temp_path) {
             File::Path::rmtree $_ if -d $_;
+          }
         }
     }
 }
